+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 2.4.1 (unreleased)
 ------------------
 
+- Fix showing private content while editing a tile.
+  Fixes `issue 100 <https://github.com/plone/plone.app.standardtiles/issues/100>`_.
+  [maurits]
+
 - Fix membertools tile
   [agitator]