+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 2.4.1 (unreleased)
 ------------------
 
+- Fixed CI: tests pass on Plonew 5.1 and 5.2, on all supported Plone versions.
+  [maurits]
+
 - Fix showing private content while editing a tile.
   Fixes `issue 100 <https://github.com/plone/plone.app.standardtiles/issues/100>`_.
   [maurits]